+++ workers/ohsome_quality_analyst/indicators/ghs_pop_comparison_buildings/indicator.py
@@ -1,12 +1,11 @@
 import logging
 from io import StringIO
 from string import Template
-from typing import Union
 
 import matplotlib.pyplot as plt
 import numpy as np
 from asyncpg import Record
-from geojson import MultiPolygon, Polygon
+from geojson import Feature
 
 from ohsome_quality_analyst.base.indicator import BaseIndicator
 from ohsome_quality_analyst.geodatabase import client as db_client
@@ -19,11 +18,11 @@ class GhsPopComparisonBuildings(BaseIndicator):
     def __init__(
         self,
         layer_name: str,
-        bpolys: Union[Polygon, MultiPolygon],
+        feature: Feature,
     ) -> None:
         super().__init__(
             layer_name=layer_name,
-            bpolys=bpolys,
+            feature=feature,
         )
         # Those attributes will be set during lifecycle of the object.
         self.pop_count = None
@@ -45,14 +44,16 @@ class GhsPopComparisonBuildings(BaseIndicator):
         return 0.75 * np.sqrt(pop_per_sqkm)
 
     async def preprocess(self) -> bool:
-        pop_count, area = await self.get_zonal_stats_population(bpolys=self.bpolys)
+        pop_count, area = await self.get_zonal_stats_population()
 
         if pop_count is None:
             pop_count = 0
         self.area = area
         self.pop_count = pop_count
 
-        query_results = await ohsome_client.query(layer=self.layer, bpolys=self.bpolys)
+        query_results = await ohsome_client.query(
+            layer=self.layer, bpolys=self.feature.geometry
+        )
         if query_results is None:
             return False
         self.feature_count = query_results["result"][0]["value"]
@@ -172,7 +173,7 @@ class GhsPopComparisonBuildings(BaseIndicator):
         plt.close("all")
         return True
 
-    async def get_zonal_stats_population(self, bpolys: dict) -> Record:
+    async def get_zonal_stats_population(self) -> Record:
         """Derive zonal population stats for given GeoJSON geometry.
 
         This is based on the Global Human Settlement Layer Population.
@@ -198,6 +199,6 @@ class GhsPopComparisonBuildings(BaseIndicator):
                 st_setsrid(public.ST_GeomFromGeoJSON($3), 4326)
              )
             """
-        data = tuple(map(str, [bpolys] * 3))
+        data = tuple(map(str, [self.feature.geometry] * 3))
         async with db_client.get_connection() as conn:
             return await conn.fetchrow(query, *data)