+++ workers/ohsome_quality_analyst/indicators/mapping_saturation/indicator.py
@@ -1,11 +1,10 @@
 import logging
 from io import StringIO
 from string import Template
-from typing import Union
 
 import matplotlib.pyplot as plt
 import pandas as pd
-from geojson import MultiPolygon, Polygon
+from geojson import Feature
 
 from ohsome_quality_analyst.base.indicator import BaseIndicator
 from ohsome_quality_analyst.indicators.mapping_saturation.sigmoid_curve import (
@@ -26,12 +25,12 @@ class MappingSaturation(BaseIndicator):
     def __init__(
         self,
         layer_name: str,
-        bpolys: Union[Polygon, MultiPolygon],
+        feature: Feature,
         time_range: str = "2008-01-01//P1M",
     ) -> None:
         super().__init__(
             layer_name=layer_name,
-            bpolys=bpolys,
+            feature=feature,
         )
         self.time_range = time_range
         # Those attributes will be set during lifecycle of the object.
@@ -42,7 +41,7 @@ class MappingSaturation(BaseIndicator):
     async def preprocess(self) -> bool:
         """Get data from ohsome API and db. Put timestamps + data in list"""
         query_results = await ohsome_client.query(
-            layer=self.layer, bpolys=self.bpolys, time=self.time_range
+            layer=self.layer, bpolys=self.feature.geometry, time=self.time_range
         )
         results = [y_dict["value"] for y_dict in query_results["result"]]
         timestamps = [y_dict["timestamp"] for y_dict in query_results["result"]]