+++ workers/ohsome_quality_analyst/reports/jrc_requirements/report.py
@@ -1,6 +1,6 @@
-from typing import Optional, Union
+from typing import Optional
 
-from geojson import MultiPolygon, Polygon
+from geojson import Feature
 
 from ohsome_quality_analyst.base.report import BaseReport, IndicatorLayer
 
@@ -8,7 +8,7 @@ from ohsome_quality_analyst.base.report import BaseReport, IndicatorLayer
 class JrcRequirements(BaseReport):
     def __init__(
         self,
-        bpolys: Union[Polygon, MultiPolygon, None] = None,
+        feature: Feature = None,
         dataset: Optional[str] = None,
         feature_id: Optional[int] = None,
         fid_field: Optional[str] = None,
@@ -16,7 +16,7 @@ class JrcRequirements(BaseReport):
         """Create a list of indicator objects."""
 
         super().__init__(
-            bpolys=bpolys, dataset=dataset, feature_id=feature_id, fid_field=fid_field
+            feature=feature, dataset=dataset, feature_id=feature_id, fid_field=fid_field
         )
 
     def set_indicator_layer(self) -> None: