+++ workers/ohsome_quality_analyst/reports/simple_report/report.py
@@ -1,6 +1,6 @@
-from typing import Optional, Union
+from typing import Optional
 
-from geojson import MultiPolygon, Polygon
+from geojson import Feature
 
 from ohsome_quality_analyst.base.report import BaseReport, IndicatorLayer
 
@@ -8,13 +8,13 @@ from ohsome_quality_analyst.base.report import BaseReport, IndicatorLayer
 class SimpleReport(BaseReport):
     def __init__(
         self,
-        bpolys: Union[Polygon, MultiPolygon] = None,
+        feature: Feature = None,
         dataset: Optional[str] = None,
         feature_id: Optional[int] = None,
         fid_field: Optional[str] = None,
     ) -> None:
         super().__init__(
-            bpolys=bpolys, dataset=dataset, feature_id=feature_id, fid_field=fid_field
+            feature=feature, dataset=dataset, feature_id=feature_id, fid_field=fid_field
         )
 
     def set_indicator_layer(self):