+++ workers/tests/integrationtests/test_api_indicator.py
@@ -30,7 +30,7 @@ class TestApiIndicator(unittest.TestCase):
         infile = os.path.join(
             os.path.dirname(os.path.abspath(__file__)),
             "fixtures",
-            "heidelberg_altstadt.geojson",
+            "heidelberg-altstadt-feature.geojson",
         )
         with open(infile, "r") as f:
             self.bpolys = geojson.load(f)