+++ workers/tests/integrationtests/test_indicator_ghs_pop_comparison_roads.py
@@ -11,13 +11,12 @@ from ohsome_quality_analyst.indicators.ghs_pop_comparison_roads.indicator import
 
 class TestIndicatorGhsPopComparisonRoads(unittest.TestCase):
     def setUp(self):
-        dataset = "regions"
-        feature_id = 3  # Heidelberg
-        bpolys = asyncio.run(
-            db_client.get_bpolys_from_db(dataset, feature_id, "ogc_fid")
+        # Heidelberg
+        feature = asyncio.run(
+            db_client.get_region_from_db(feature_id=3, fid_field="ogc_fid")
         )
         self.indicator = GhsPopComparisonRoads(
-            bpolys=bpolys, layer_name="major_roads_length"
+            feature=feature, layer_name="major_roads_length"
         )
 
     def test(self):
@@ -37,9 +36,7 @@ class TestIndicatorGhsPopComparisonRoads(unittest.TestCase):
         self.assertIsNotNone(self.indicator.result.svg)
 
     def test_get_zonal_stats_population(self):
-        result = asyncio.run(
-            self.indicator.get_zonal_stats_population(self.indicator.bpolys)
-        )
+        result = asyncio.run(self.indicator.get_zonal_stats_population())
         self.assertIsInstance(result, Record)