+++ workers/tests/integrationtests/test_oqt.py
@@ -9,20 +9,23 @@ from .utils import oqt_vcr
 
 class TestOqt(unittest.TestCase):
     def setUp(self):
+        # Heidelberg
         self.dataset = "regions"
-        self.feature_id = 3  # Heidelberg
+        self.feature_id = 3
         self.fid_field = "ogc_fid"
-        self.bpolys = asyncio.run(
-            db_client.get_bpolys_from_db(self.dataset, self.feature_id, self.fid_field)
+        self.feature = asyncio.run(
+            db_client.get_region_from_db(
+                feature_id=self.feature_id, fid_field=self.fid_field
+            )
         )
 
     @oqt_vcr.use_cassette()
-    def test_create_indicator_bpolys(self):
+    def test_create_indicator_feature(self):
         """Test creating indicator from scratch."""
         # From scratch
         indicator = asyncio.run(
             oqt.create_indicator(
-                "GhsPopComparisonBuildings", "building_count", bpolys=self.bpolys
+                "GhsPopComparisonBuildings", "building_count", feature=self.feature
             )
         )
         self.assertIsNotNone(indicator.result.label)
@@ -111,13 +114,13 @@ class TestOqt(unittest.TestCase):
                     "building_count",
                     dataset=self.dataset,
                     feature_id=self.feature_id,
-                    bpolys=self.bpolys,
+                    feature=self.feature,
                 )
             )
 
     @oqt_vcr.use_cassette()
-    def test_create_report_bpolys(self):
-        report = asyncio.run(oqt.create_report("SimpleReport", bpolys=self.bpolys))
+    def test_create_report_feature(self):
+        report = asyncio.run(oqt.create_report("SimpleReport", feature=self.feature))
         self.assertIsNotNone(report.result.label)
         self.assertIsNotNone(report.result.value)
         self.assertIsNotNone(report.result.description)